Split devtools utils
